requirements.txt
# requirements
requests
python-dotenv

# dev requirements
pytest
